requirements.txt
pandas
pyarrow
scikit-surprise
joblib
flask
boto3